moved and split up schemas awaiting comments/advice
